Handle null value and non existant value
In our project i have handled saving the data to a storage driver myself (added it well before it was included in the current package). 

When adding a test for this functionality in my project the latest version of the code throws an error about attempting to get a value from a null. 

I've seen you check for the values presence in other methods in this class, so i added it here too.